pyproject.toml
[project]
name = "ask-question-get-answer-api"
version = "0.1.0"
description = ""
authors = [
    {name = "rukiqq",email = "mitaka@duck.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.121.1,<0.122.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "fastapi-users[sqlalchemy] (>=15.0.1,<16.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.0,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
